ar_module/requirements.txt
torch~=2.7.1+cu126
transformers~=4.52.4
numpy~=2.1.2
opencv-python~=4.11.0.86
pillow~=11.0.0
trimesh~=4.6.12
scipy~=1.15.3
pyrender~=0.1.45
open3d~=0.19.0